convert slot-config to typescript
