+++ static/src/javascripts/projects/commercial/modules/header-bidding/types.d.ts
@@ -119,3 +119,5 @@ type PrebidMediaTypes = {
 		sizes: HeaderBiddingSize[];
 	};
 };
+
+type SlotFlatMap = (slot: HeaderBiddingSlot) => HeaderBiddingSlot[];
